01-research-scraper/coding_agents/copilot_opus-45/pyproject.toml
[project]
name = "researcher"
version = "0.1.0"
description = "PaperTrail - Research Paper Catalog with GraphRAG-based search"
readme = "README.md"
authors = [
    { name = "Pyro", email = "pyros.sd.models@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "litellm>=1.50.0",
    "sentence-transformers>=3.3.0",
    "tinydb>=4.8.0",
    "networkx>=3.4.0",
    "numpy>=2.0.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "websockets>=14.0",
    "feedparser>=6.0.0",
    "arxiv>=2.1.0",
    "pypdf>=5.0.0",
    "pydantic>=2.10.0",
    "python-multipart>=0.0.17",
]

[project.scripts]
researcher = "researcher:main"

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]
